+++ app/pages/articles/[id]/edit/ArticleEdit.tsx
@@ -3,11 +3,14 @@
 import { Footer } from '@/components/elements/Footer'
 import styles from '@/components/layouts/SiteWrapper/index.module.css'
 import { StickyWrapper } from '@/components/layouts/StickyWrapper'
+import { useLens } from '@hookform/lenses'
+import { zodResolver } from '@hookform/resolvers/zod'
 import { FC } from 'react'
+import { FormProvider, useForm } from 'react-hook-form'
 import { Article, ArticleCategory } from '../../shared/article.types'
 import { ArticleForm } from '../../shared/articleForm/ArticleForm'
+import { Schema, schema } from '../../shared/articleForm/articleForm.schema'
 import { ArticleEditHeader } from './components/articleEditHeader/ArticleEditHeader'
-import { ArticleEditProvider } from './components/articleEditProvider/ArticleEditProvider'
 
 const BLOCK_NAME = 'site-wrapper'
 
@@ -19,17 +22,35 @@ type ArticleEditProps = {
 export const ArticleEdit: FC<ArticleEditProps> = ({ article, categories }) => {
   const categoryOptions = categories?.map((category) => category.name) ?? []
 
+  const defaultValues = {
+    title: article?.title ?? '',
+    content: article?.content ?? '',
+    categories:
+      article?.categories?.map(({ category }) => ({ name: category.name })) ??
+      [],
+    status: (article?.status ?? 'PUBLISHED') as Schema['status'],
+  }
+
+  const methods = useForm<Schema>({
+    mode: 'onTouched',
+    resolver: zodResolver(schema),
+    defaultValues,
+  })
+
+  const lens = useLens<Schema>({ control: methods.control })
+
+  // NOTE: useLensを使うことでArticleFormがFormProviderの外にあっても動作するようになる
   return (
-    <ArticleEditProvider article={article}>
-      <div className={styles[`${BLOCK_NAME}`]}>
-        <StickyWrapper>
+    <div className={styles[`${BLOCK_NAME}`]}>
+      <StickyWrapper>
+        <FormProvider {...methods}>
           <ArticleEditHeader article={article} />
-        </StickyWrapper>
-        <main className={styles[`${BLOCK_NAME}-inner`]}>
-          <ArticleForm categoryOptions={categoryOptions} />
-        </main>
-        <Footer />
-      </div>
-    </ArticleEditProvider>
+        </FormProvider>
+      </StickyWrapper>
+      <main className={styles[`${BLOCK_NAME}-inner`]}>
+        <ArticleForm lens={lens} categoryOptions={categoryOptions} />
+      </main>
+      <Footer />
+    </div>
   )
 }
